Feat; Saving DB with device hostname
