+++ src-go/browser/info.go
@@ -19,7 +19,6 @@ const (
 )
 
 var availableBrowsers = []Browser{
-	GetTesterBrowser(),
 	GetFirefoxBrowser(),
 	GetFirefoxDeveloperBrowser(),
 	GetChromiumBrowser(),
